+++ hdt-lib/src/triples/TripleListDisk.cpp
@@ -239,7 +239,7 @@ void TripleListDisk::populateHeader(Header &header, string rootNode)
 	header.insert(rootNode, HDTVocabulary::TRIPLES_ORDER, order );  // TODO: Convert to String
 }
 
-unsigned int TripleListDisk::getNumberOfElements()
+size_t TripleListDisk::getNumberOfElements()
 {
 	return numValidTriples;
 }
@@ -343,7 +343,7 @@ bool TripleListDisk::remove(IteratorTripleID *triples)
 
 	bool removed = false;
 	for(tid=arrayTriples; tid<arrayTriples+numTotalTriples; tid++) {
-		for(int i=0; i<allPat.size(); i++) {
+		for(size_t i=0; i<allPat.size(); i++) {
 			if (tid->match(allPat[i])) {
 				tid->clear();
 				numValidTriples--;
